grantio-scraper/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "grants-scraper"
version = "0.1.0"
description = "Czech grants scraper with modular architecture"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Grantio", email = "dev@grantio.cz" }
]
keywords = ["grants", "scraping", "czech", "dotace"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "structlog>=24.0",
    "tenacity>=8.0",
    "python-dateutil>=2.8",
]

[project.optional-dependencies]
pdf = [
    "pdfplumber>=0.11.0",
]
excel = [
    "openpyxl>=3.1.0",
    "pandas>=2.0",
]
llm = [
    "openai>=1.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "mypy>=1.8",
    "ruff>=0.3",
]
all = [
    "grants-scraper[pdf,excel,llm,dev]",
]

[project.scripts]
grants-scraper = "grants_scraper.__main__:main"

[project.urls]
Homepage = "https://github.com/grantio/grants-scraper"
Repository = "https://github.com/grantio/grants-scraper"

[tool.hatch.build.targets.wheel]
packages = ["grants_scraper"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
